fixed alias expansion
